refactor organization skill manage helpers
